Track every retention, split the screen's mobile cases, scope receipt metadata
Retentions for one Card can overlap, so each future is tracked until it
settles. The four mobile cases of the Gift Cards screen test move into a
@Tags(['mobile']) file with the shared harness in test/support. The
desktop receipt uses passed-in Gift Card metadata only under the account
it was resolved for.
