+++ lib/src/features/payment_links/providers/payment_link_claim_coordinator_provider.dart
@@ -50,7 +50,7 @@ class PaymentLinkClaimCoordinator {
 
   final Ref _ref;
   final Map<String, Future<PaymentLinkClaimResult>> _submissions = {};
-  final Map<String, Future<void>> _retentions = {};
+  final Set<Future<void>> _retentions = {};
   Future<List<PaymentLinkReceivedRecord>>? _recoveryInFlight;
   Timer? _retryTimer;
   bool _enabled = false;
@@ -90,20 +90,20 @@ class PaymentLinkClaimCoordinator {
 
   /// A retention writes the received store, so a reset drains it; one started
   /// after quiesce is skipped so it cannot resurrect a Card in the wiped wallet.
-  Future<void> trackRetention(String address, Future<void> Function() run) {
+  Future<void> trackRetention(Future<void> Function() run) {
     if (_resetQuiesced) {
       debugPrint(
         '[zcash] PaymentLinkClaim: skipped retaining a claim during reset',
       );
       return Future<void>.value();
     }
+    // Every retention is tracked on its own: retentions for one Card can
+    // overlap, and a reset has to outlast the slowest of them.
     late final Future<void> tracked;
-    tracked = Future<void>.sync(run).whenComplete(() {
-      if (identical(_retentions[address], tracked)) {
-        _retentions.remove(address);
-      }
-    });
-    _retentions[address] = tracked;
+    tracked = Future<void>.sync(
+      run,
+    ).whenComplete(() => _retentions.remove(tracked));
+    _retentions.add(tracked);
     return tracked;
   }
 
@@ -148,7 +148,7 @@ class PaymentLinkClaimCoordinator {
         _recoveryInFlight != null) {
       final pending = <Future<Object?>>[
         ..._submissions.values,
-        ..._retentions.values,
+        ..._retentions,
         ?_recoveryInFlight,
       ];
       await Future.wait(pending.map(_ignoreOutcome));
